+++ deps.ts
@@ -1,3 +1,3 @@
-export * from "https://deno.land/std@0.128.0/textproto/mod.ts";
-export * from "https://deno.land/std@0.128.0/async/mod.ts";
-export * from "https://deno.land/std@0.128.0/io/mod.ts";
+export * from "https://deno.land/std@0.129.0/textproto/mod.ts";
+export * from "https://deno.land/std@0.129.0/async/mod.ts";
+export * from "https://deno.land/std@0.129.0/io/mod.ts";